fix(#463): apply fix pass changes
